requirements.txt
future==0.18.0
imbalanced-learn==0.5.0
matplotlib==3.3.4
networkx==2.3
numpy==1.17.2
openml==0.9.0
pandas==0.25.1
prettytable==0.7.2
python-dateutil==2.8.0
PyYAML==5.1.2
scikit-learn==0.22
scipy==1.3.1
s-dbw==0.4.0
tabulate==0.8.9
tqdm==4.36.1
plotly==5.15.0
kaleido==0.2.1
